[MOV] web, website: move still used public legacy code
