+++ addons/web/static/src/public/public_root_instance.js
@@ -1,4 +1,4 @@
 import { createPublicRoot } from "./public_root";
-import lazyloader from "@web/legacy/js/public/lazyloader";
+import lazyloader from "@web/public/lazyloader";
 
 lazyloader.registerPageReadinessDelay(createPublicRoot());
