+++ addons/website/__manifest__.py
@@ -247,7 +247,7 @@
             # Activated on-demand by website.ripple_effect_js.
             ('remove', 'website/static/src/interactions/ripple_effect.js'),
             ('remove', 'website/static/src/core/website_edit_service.js'),
-            ('replace', 'web/static/src/legacy/js/public/public_root_instance.js', 'website/static/src/js/content/website_root_instance.js'),
+            ('replace', 'web/static/src/public/public_root_instance.js', 'website/static/src/js/content/website_root_instance.js'),
             'website/static/src/snippets/**/*.js',
             ('remove', 'website/static/src/snippets/**/*.edit.js'),
             'website/static/src/libs/zoomodoo/zoomodoo.scss',
@@ -368,8 +368,8 @@
             'html_builder/static/src/utils/scrolling.js',
             'web/static/src/legacy/js/core/class.js',
             'web/static/src/public/utils.js',
-            'web/static/src/legacy/js/public/lazyloader.js',
-            'web/static/src/legacy/js/public/public_root.js',
+            'web/static/src/public/lazyloader.js',
+            'web/static/src/public/public_root.js',
             'website/static/lib/multirange/*.js',
             'website/static/src/js/content/auto_hide_menu.js',
             'website/static/src/core/**/*',
